Moved uploading image logic from CameraCapture to its own file
